tools/requirements.txt
agno==1.0.8
docker_py==1.10.6
docx2pdf==0.1.8
llama_index==0.12.19
nest_asyncio==1.6.0
openai==0.28.0
protobuf==5.29.3
psutil==6.1.1
python-dotenv==1.0.1
python_docx==1.1.2
